add cross segement mss tests

+++ lib/common/mss.cpp
@@ -11,26 +11,26 @@ void char_buffer_init(charBuffer *cb, char uf, char d, char of) {
   cb->overflow[7]  = 0;
 }
 
-void update_by_index(charBuffer& cb, int offset, int size, int step, char c) {
-  for(int i=offset; i != size+offset; i += step)
+void update_by_index(charBuffer& cb, long long offset, long long size, int step, char c) {
+  for(long long i=offset; i != size+offset; i += step)
     cb.data[i] = c;
 }
 
-void update_by_pointer(char *buf, int offset, int size, int step, char c) {
+void update_by_pointer(char *buf, long long offset, long long size, int step, char c) {
   buf += offset;
-  for(int i=0; i != size; i += step, buf += step)
+  for(long long i=0; i != size; i += step, buf += step)
     *buf = c;
 }
 
-int read_by_index(const charBuffer& cb, int offset, int size, int step, char c) {
-  for(int i=offset; i != size+offset; i += step)
+int read_by_index(const charBuffer& cb, long long offset, long long size, int step, char c) {
+  for(long long i=offset; i != size+offset; i += step)
     if(cb.data[i] != c) return 1;
   return 0;
 }
 
-int read_by_pointer(const char *buf, int offset, int size, int step, char c) {
+int read_by_pointer(const char *buf, long long offset, long long size, int step, char c) {
   buf += offset;
-  for(int i=0; i != size; i += step, buf += step)
+  for(long long i=0; i != size; i += step, buf += step)
     if(*buf != c) return 1;
   return 0;
 }
