+++ lib/include/mss.hpp
@@ -9,10 +9,10 @@ struct charBuffer
 };
 
 extern void char_buffer_init(charBuffer *, char uf, char d, char of);
-extern void update_by_index(charBuffer& cb, int offset, int size, int step, char c);
-extern void update_by_pointer(char *buf, int offset, int size, int step, char c);
-extern int read_by_index(const charBuffer& cb, int offset, int size, int step, char c);
-extern int read_by_pointer(const char *buf, int offset, int size, int step, char c);
+extern void update_by_index(charBuffer& cb, long long offset, long long size, int step, char c);
+extern void update_by_pointer(char *buf, long long offset, long long size, int step, char c);
+extern int read_by_index(const charBuffer& cb, long long offset, long long size, int step, char c);
+extern int read_by_pointer(const char *buf, long long offset, long long size, int step, char c);
 extern int check(const char *buf, int size, int step, char c);
 
 #endif
